requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
cloudinary==1.44.0
Django==5.2
django-cloudinary-storage==0.3.0
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0

/var/www/devnahid_pythonanywhere_com_wsgi.py


